+++ src/main/java/ca/kpu/info2413/library/backend/service/BookCopyService.java
@@ -46,7 +46,7 @@ public class BookCopyService
     }
 
     // Find by ISBN13
-    public List<BookCopy> findByIsbn13(Integer isbn13)
+    public List<BookCopy> findByIsbn13(Long isbn13)
     {
         Optional<Publication> publicationOpt = publicationRepository.findById(isbn13);
         if (publicationOpt.isEmpty())
